FIXES AND NEWS
Fixing bug on /_layout for login/home redirection depending on auth state. Missing the [auth] condition in the useEffect, without subscribing to onAuthStateChanged.
Fixing bug on packages/app/features/auth/login for missing [auth] condition in the useEffect for redirecting to home
Adding redirection to login  when not logged in at each route: /(auth)/documents|home|profile
Adding redirection to login when not signed in at /
Adding logout feature at the Profile screen.