Update
Removing authLayout
Adding next routes to have the same routes on mobile.
Adding a siwtch that can be hidden if displayed on a phone.
   Because using expo-device in web (next,solito) was causing problems,
   a component for each plattform was created.
Changing redirection links from '/(auth)/login' to '/login', according
   to routes documentation on both, expo and next.
tailwind only works on nextjs but not on the packages/app/features/* screens.

User/Password Authentication OK