BUG FIX
Changing to mixed inlie style with StyleSheet.create() for fixing the bug where
no colors defined by variables, padding/margin or font styles where not applied.

The fix was remove the varialbes from the StyleSheet.create() because it is not
recreated when variables change. Also, when changing padding/margins or font styles
where not applied when those styles where in the StyleSheet.create(), so were moved
to inline style.